+++ internal/persistence/model/user_test.go
@@ -12,7 +12,7 @@ func TestNewUser(t *testing.T) {
 	u := NewUser(0, "johndoe", "John Doe", "john@example.com", birthDate, "123456789")
 
 	assert.NotNil(t, u)
-	assert.Equal(t, int64(0), u.UserId)
+	assert.Equal(t, int32(0), u.UserId)
 	assert.Equal(t, "johndoe", u.Username)
 	assert.Equal(t, "John Doe", u.Fullname)
 	assert.Equal(t, "john@example.com", u.Email)
